pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
flask = "^3.0.0"
gunicorn = "^21.2.0"
alembic = "^1.13.1"
Flask-SQLAlchemy = "^3.1.1"
openai = "^1.43.0"
pandas = "^2.2.3"
openpyxl = "^3.1.2"  # Add this line
sqlalchemy = "^2.0.36"
tenacity = "^9.0.0"
authlib = "^1.3.2"
zsvision = "^0.7.12"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"
# psycopg2 = "^2.9"
psycopg2-binary = "^2.9.10"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"